feat: add app.js file for non-bin-hoisted environments
